Remove SDF shader

+++ src/graphics/WebGpuRenderer/RenderEncoder.hpp
@@ -39,7 +39,7 @@ public:
     void setVisualSettings(const VisualSettings& visualSettings) final;
 
     void begin(Rc<RenderTarget> target, std::optional<ColorF> clear = Palette::transparent) final;
-    void batch(std::span<const RenderState> commands, std::span<const float> data) final;
+    void batch(std::span<const RenderState> commands, std::span<const uint32_t> data) final;
     void end() final;
     void wait() final;
     Rc<RenderTarget> currentTarget() const;
@@ -95,7 +95,7 @@ private:
 
     wgpu::BindGroup createBindGroup(ImageBackendWebGpu* imageBackend);
     void updatePerFrameConstantBuffer(const ConstantPerFrame& constants);
-    void updateDataBuffer(std::span<const float> data);
+    void updateDataBuffer(std::span<const uint32_t> data);
     void updateConstantBuffer(std::span<const RenderState> data);
     void updateAtlasTexture();
     void updateGradientTexture();
